requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
